@@ -3,6 +3,7 @@
 
 #include <set>
 #include <stddef.h>
+#include <memory>
 
 #define KNOWN_PRIMES_LIMIT 104729
 
@@ -10,14 +11,13 @@ class known_primes
 {
 public:
     known_primes();
-    virtual ~known_primes();
 
     std::set<size_t>& set()
     {
         return *known_primes_;
     }
 private:
-    std::set<size_t>* known_primes_;
+    std::unique_ptr< std::set<size_t> > known_primes_;
 };
 
 #endif // KNOWN_PRIMES_HPP_
